+++ certbot-compatibility-test/certbot_compatibility_test/configurators/common.py
@@ -4,7 +4,7 @@ import os
 import shutil
 import tempfile
 
-from certbot import constants
+from certbot._internal import constants
 from certbot_compatibility_test import errors
 from certbot_compatibility_test import util
