+++ certbot/tests/account_test.py
@@ -278,7 +278,7 @@ class AccountFileStorageTest(test_util.ConfigTestCase):
         self._set_server('https://acme-staging.api.letsencrypt.org/directory')
         self.storage.save(self.acc, self.mock_client)
         # ensure v2 isn't already linked to it
-        with mock.patch('certbot.constants.LE_REUSE_SERVERS', {}):
+        with mock.patch('certbot._internal.constants.LE_REUSE_SERVERS', {}):
             self._set_server('https://acme-staging-v02.api.letsencrypt.org/directory')
             self.assertRaises(errors.AccountNotFound, self.storage.load, self.acc.id)
