+++ certbot/util.py
@@ -19,7 +19,7 @@ import six
 
 from acme.magic_typing import Tuple, Union  # pylint: disable=unused-import, no-name-in-module
 
-from certbot import constants
+from certbot._internal import constants
 from certbot import errors
 from certbot._internal import lock
 from certbot.compat import os
